first-penguin-back/api/requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
jwt==1.0.0
oauthlib==3.1.0
pkg-resources==0.0.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
uWSGI==2.0.19.1
